feat: moved token parsing out of the client (#9)
Parsing of the Discord bot token was being done directly by the client, which needs the token to start. This led to weird behavior where the client would read the token, the application would then retrieve it from the client to then pass back into the client's `run` method. Rather than doing that, we put the onus of reading the token on the application instead.