feat: updated bot to refresh its rules at a regular cadence (#10)
The bot was modified to automatically refresh its rules at a regular cadence. The current logic is that, after one minute, the bot will check the modification time of the rules file every time a message is sent on the server, and if has changed, the rules will be updated.

The main bot entry point was modified to extend the `commands.Bot` class, rather than the `discord.Client`. This will give us access to commands in the future.